products helper refacto + fix front show product
