+++ app/views/pages/home.html.erb
@@ -3,7 +3,7 @@
     <h1 class="pb-3 pb-xl-5">Goûts d’Fruits : la ferme fruitière de Marie Gazeau</h1>
     <p class="pt-3 pt-xl-5 pb-2 pb-xl-4">vous propose fruits bios de saison, pépinière et de nombreuses douceurs.</p>
     <p class="pt-3 pt-xl-5 pb-2 pb-xl-5">Des fruits frais aux desserts gourmands, tout est produit à la ferme et vendu en circuits courts.</p>
-    <%= link_to new_client_session_path do %>
+    <%= link_to products_path do %>
       <button type="button" class="btn submit-btn-green-no-width middle-text-resp bold text-wrap">Commander</button>
     <% end %>
   </div>
