+++ app/views/products/show.html.erb
@@ -10,11 +10,18 @@
             <i class="far fa-trash-alt blue big-text"></i>
           <% end %>
         </div>
+        <p class="card-product-show-title text-center"><%=@product.name.capitalize%></p>
         <div class="card-product-show-flex">
-          <p class="card-product-show-title"><%=@product.name.capitalize%></p>
           <span class="card-product-show-price">
-          <%= @product.display_price(current_client, @type_price) %> €
-          <sup>/ <%= @product.unit_type %> </sup></span>
+            <%= display_price(@product, current_client, 'non-magasin') %> €
+            <sup>/ <%= @product.unit_type %>
+            de <%= @product.unit_measure_quantity %> <%= @product.unit_measure %></sup>
+          </span>
+          <span class="card-product-show-price">
+            <%= display_price(@product, current_client, 'magasin') %> €
+            <sup>/ <%= @product.unit_type %>
+            de <%= @product.unit_measure_quantity_shop %> <%= @product.unit_measure %></sup>
+          </span>
         </div>
         <p class="product-show-description pt-4"><%= @product.description%></p>
